refactor into checks dir
