+++ tools/checks/mandatory.js
